Rename observeHelpers.js to observe-helpers.js
